Fixed castling * Now able to castle * Fixed Square type requiring cast issue => Still need to make castling more rubost and cancel the move

+++ tests/move/tests_play_move.cpp
@@ -13,7 +13,7 @@ TEST_F(MovePlayTest, Play)
 {
     Board b{};
     b.load_fen("8/1p1q4/5k2/1n1R3r/8/1K2N1p1/3B4/8 w - - 0 1");
-    Move move(static_cast<int>(Square::d5), static_cast<int>(Square::d6), ROOK);
+    Move move(Square::d5, Square::d6, ROOK);
 
     b.play(move);
     ASSERT_EQ(b.get_occupancy(NO_COLOR), 2859288583342080);
@@ -23,7 +23,7 @@ TEST_F(MovePlayTest, PlayUnplay)
 {
     Board b{};
     b.load_fen("8/1p1q4/5k2/1n1R3r/8/1K2N1p1/3B4/8 w - - 0 1");
-    Move move(static_cast<int>(Square::d5), static_cast<int>(Square::d6), ROOK);
+    Move move(Square::d5, Square::d6, ROOK);
     const bitboard occ_in = b.get_occupancy(NO_COLOR);
     const bitboard occ_white_in = b.get_occupancy(WHITE);
     const bitboard occ_white_rook_in = b.get_piece_bitboard(WHITE, ROOK);
@@ -43,7 +43,7 @@ TEST_F(MovePlayTest, PlayUnplayCapture)
 {
     Board b{};
     b.load_fen("8/1p1q4/5k2/1n1R3r/8/1K2N1p1/3B4/8 w - - 0 1");
-    Move move(static_cast<int>(Square::d5), static_cast<int>(Square::d7), ROOK);
+    Move move(Square::d5, Square::d7, ROOK);
     const bitboard occ_in = b.get_occupancy(NO_COLOR);
     const bitboard occ_white_in = b.get_occupancy(WHITE);
     const bitboard occ_white_rook_in = b.get_piece_bitboard(WHITE, ROOK);
